+++ src/main/java/net/frozenblock/lib/block/mixin/fire/AbstractArrowMixin.java
@@ -17,7 +17,11 @@
 
 package net.frozenblock.lib.block.mixin.fire;
 
+import java.util.Optional;
+import net.frozenblock.lib.block.api.fire.FireEvents;
 import net.frozenblock.lib.block.impl.fire.FireData;
+import net.frozenblock.lib.block.impl.fire.FireType;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
 import net.minecraft.world.phys.EntityHitResult;
 import org.spongepowered.asm.mixin.Mixin;
@@ -36,9 +40,12 @@ public class AbstractArrowMixin {
 		)
 	)
 	public void frozenLib$setFireType(EntityHitResult hitResult, CallbackInfo info) {
-		final FireData fireData = AbstractArrow.class.cast(this).getAttached(FireData.ATTACHMENT);
-		if (fireData == null) return;
-
-		FireData.trySet(hitResult.getEntity(), fireData.type());
+		final ResourceKey<FireType> fireType = FireEvents.SELECT_FIRE_TYPE.invoker().selectFireType(
+			hitResult.getEntity(),
+			Optional.empty(),
+			Optional.of(AbstractArrow.class.cast(this)),
+			Optional.empty()
+		);
+		FireData.trySet(hitResult.getEntity(), fireType);
 	}
 }