+++ src/main/java/net/frozenblock/lib/block/mixin/fire/BaseFireBlockMixin.java
@@ -17,26 +17,44 @@
 
 package net.frozenblock.lib.block.mixin.fire;
 
-import net.frozenblock.lib.block.api.fire.FireTypes;
+import java.util.Optional;
+import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import com.llamalad7.mixinextras.sugar.Local;
+import net.frozenblock.lib.block.api.fire.FireEvents;
 import net.frozenblock.lib.block.impl.fire.FireData;
 import net.frozenblock.lib.block.impl.fire.FireType;
-import net.minecraft.core.Holder;
+import net.minecraft.core.BlockPos;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.entity.Entity;
+import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.block.BaseFireBlock;
+import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-import java.util.Optional;
 
 @Mixin(BaseFireBlock.class)
 public class BaseFireBlockMixin {
 
 	@Inject(method = "lambda$entityInside$0", at = @At("HEAD"))
 	public void frozenLib$setFireType(Entity entity, CallbackInfo info) {
-		final Optional<Holder<FireType>> fireType = FireTypes.getTypeForBlock(entity.registryAccess(), BaseFireBlock.class.cast(this));
-		if (fireType.isEmpty()) return;
+		final ResourceKey<FireType> fireType = FireEvents.SELECT_FIRE_TYPE.invoker().selectFireType(
+			entity,
+			Optional.of(BaseFireBlock.class.cast(this)),
+			Optional.empty(),
+			Optional.empty()
+		);
+		FireData.trySet(entity, fireType);
+	}
 
-		FireData.trySet(entity, fireType.get());
+	@ModifyReturnValue(method = "getState", at = @At("RETURN"))
+	private static BlockState frozenLib$selectFireBlockState(
+		BlockState original,
+		@Local(argsOnly = true) BlockGetter level,
+		@Local(name = "below") BlockPos below, @Local(name = "belowState") BlockState belowState
+	) {
+		final BlockState newState = FireEvents.SELECT_FIRE_BLOCK_STATE.invoker().selectFireBlockState(level, below, belowState);
+		return newState != null ? newState : original;
 	}
 }