+++ src/main/java/net/frozenblock/lib/block/mixin/fire/WitherSkeletonMixin.java
@@ -18,7 +18,11 @@
 package net.frozenblock.lib.block.mixin.fire;
 
 import com.llamalad7.mixinextras.sugar.Local;
+import java.util.Optional;
+import net.frozenblock.lib.block.api.fire.FireEvents;
 import net.frozenblock.lib.block.impl.fire.FireData;
+import net.frozenblock.lib.block.impl.fire.FireType;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.entity.monster.skeleton.WitherSkeleton;
 import net.minecraft.world.entity.projectile.arrow.AbstractArrow;
 import net.minecraft.world.item.ItemStack;
@@ -44,6 +48,12 @@ public class WitherSkeletonMixin {
 		final FireData fireData = WitherSkeleton.class.cast(this).getAttached(FireData.ATTACHMENT);
 		if (fireData == null || !fireData.type().value().spreadsFromIgniteEnchantments()) return;
 
-		FireData.trySet(arrow, fireData.type());
+		final ResourceKey<FireType> fireType = FireEvents.SELECT_FIRE_TYPE.invoker().selectFireType(
+			arrow,
+			Optional.empty(),
+			Optional.of(WitherSkeleton.class.cast(this)),
+			Optional.empty()
+		);
+		FireData.trySet(arrow, fireType);
 	}
 }