+++ src/main/java/net/frozenblock/lib/block/mixin/fire/ZombieMixin.java
@@ -17,7 +17,11 @@
 
 package net.frozenblock.lib.block.mixin.fire;
 
+import java.util.Optional;
+import net.frozenblock.lib.block.api.fire.FireEvents;
 import net.frozenblock.lib.block.impl.fire.FireData;
+import net.frozenblock.lib.block.impl.fire.FireType;
+import net.minecraft.resources.ResourceKey;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.monster.zombie.Zombie;
@@ -40,6 +44,12 @@ public class ZombieMixin {
 		final FireData fireData = Zombie.class.cast(this).getAttached(FireData.ATTACHMENT);
 		if (fireData == null || !fireData.type().value().spreadsFromZombie()) return;
 
-		FireData.trySet(target, fireData.type());
+		final ResourceKey<FireType> fireType = FireEvents.SELECT_FIRE_TYPE.invoker().selectFireType(
+			target,
+			Optional.empty(),
+			Optional.of(Zombie.class.cast(this)),
+			Optional.empty()
+		);
+		FireData.trySet(target, fireType);
 	}
 }